Updates FeatureIdPipelineStageSpec
